pies/pi3/pyproject.toml
[project]
name = "pi3"
version = "2026.2.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "gpiohero",
    "paho-mqtt>=2.1.0",
]
